Move send_error() method to cogs._utils
